Corrections to tutorial with further testing and fixes
I discovered that a recent change was causing the pipeline loss to
report 0. The issue was the recent change, which adds a context manager
to set the default torch datatype for the entire training session.

Specifically, the computed loss on the last stage was explicitly cast to
float32, while the other stages created an empty, default, tensor to
sync via broadcast. When the default type of that tensor became
bfloat16, the broadcast no longer sync'd the loss.

I fixed that bug and tracked down a number of other places where similar
issues may occur (optimizers). This resolved the issue.

Ultimately, I decided to narrow the scope of the default type context
manager to just model construction. Allowing this to apply to everything
is begging to cause strange issues in code outside of Forgather. This
should be safer.

Other changes:
- Added a 4 GPU float32 pipeline training configuraiton.
- Added (disabled) instrumentation to check grad-norms of all tensors.
  Useful for debugging.
- Updated config syntax on 1f1b_4gpu.yaml, as I missed it on the last
  pass.
- Disabled validation on the zbvz configs, as it's broken.
- Added configs for AdamW and AdamW-4bit optimizers.
- Added code to the data-collator to check and fix the padding
  alignment. Some tokenizers are configured to pad on the left. As we
don't pass an attention mask, this produces incorrect results, so we
move padding/truncation to the right.

+++ src/forgather/ml/optim/adafactor.py
@@ -39,7 +39,7 @@ class Adafactor(Optimizer):
         super().__init__(params, defaults)
 
     def _init_state(self, state, group, p, grad):
-        state["step"] = torch.tensor(0.0)
+        state["step"] = torch.tensor(0.0, dtype=torch.float32)
         if grad.dim() <= 1:
             state["row"] = torch.zeros_like(grad, dtype=p.dtype)
             state["col"] = None
